Refactor firebase auth code
