+++ src/app/(auth)/authenticated-route.tsx
