+++ src/app/(auth)/signup/page.tsx
@@ -1,6 +1,6 @@
 import SignupForm from "./signup-form";
 
-export default function LoginPage() {
+export default function SignupPage() {
   return (
     <main className="flex h-full w-full items-center justify-center">
       <SignupForm />
